Updated GRIS HTML and CSS to fix visual display
Added class .wide--grid to layout containers so that the GRID is fits to the full area available and the Keyboard instructions is block level underneath.
Removed the absolute positioning of the 'Programmatically update the value...' button to sit under the radio buttons allowing for a narrower screen width.

Note:  the GRIDs breaks when screens fall below 1020px wide. Need to investigate reflow options which will take a bit of work.