Temp solution apply horizontal scroll to DIV if screen < 65em wide
Bryan, this is a temporary solution which adds horizontal scrollbars when the grid is reduced to a widths where the content overlaps.
I'll work on an improved solution as time allows. At least the GRID no linger breaks on mobile and screens < 1000px wide.

I also made one small change the layout CSS to fix an issue I discovered.